ai-image-project/worker-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
boto3==1.36.7
Pillow==11.1.0
